+++ sdmap/test/sdmap.unittest/LexerTest/SqlTextTest.cs
@@ -1,0 +1,48 @@
+﻿using Antlr4.Runtime;
+using sdmap.Parser.G4;
+using sdmap.Utils;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Xunit;
+
+using static sdmap.Parser.G4.SdmapLexer;
+
+namespace sdmap.unittest.LexerTest
+{
+    public class SqlTextTest
+    {
+        [Fact]
+        public void DoubleHashIsHash()
+        {
+            var code = "sql v1{##}";
+            var ats = new AntlrInputStream(code);
+            var lexer = new SdmapLexer(ats);
+            var tokens = lexer.GetAllTokens();
+            Assert.Equal(
+                new[] { KSql, SYNTAX, OpenCurlyBrace, SQLText, CloseSql },
+                tokens.Select(x => x.Type));
+        }
+
+        [Fact]
+        public void SingleHashIsMacro()
+        {
+            var code = "sql v1{#}";
+            var ats = new AntlrInputStream(code);
+            var lexer = new SdmapLexer(ats);
+            var tokens = lexer.GetAllTokens();
+            Assert.Equal(
+                new[] { KSql, SYNTAX, OpenCurlyBrace, Hash, CloseCurlyBrace },
+                tokens.Select(x => x.Type));
+        }
+
+        [Fact]
+        public void DoubleHashWillEmitSingleHash()
+        {
+            var sqlText = "##";
+            var result = SqlTextUtil.Parse(sqlText);
+            Assert.Equal("#", result);
+        }
+    }
+}